frontend: add timer view for EthCC
This is a special feature only for the EthCC convention. It got added within the
frontend to develop very quickly as all the setup and surroundings are already
available.
This adds a new view available under `/timer`. It will continuously listen for
new requests, start a timer and stops the timer when the request got filled.
Then it waits until the next request. Intermediate requests get ignored.

**ATTENTION:**
This is a hacked feature without any proper engineering. The code is ugly as
hell and should never be merged or used in production! Please throw away after
usage.